refactor(shared-research-output): extract header card component
